requirements.txt
Flask==2.3.3
Flask-Login==0.6.2
pymongo==4.5.0
gridfs==2.1.0
groq==0.1.0  # Please check the exact version of the Groq package.
werkzeug==2.3.6
python-dotenv==1.0.0
requests==2.28.2